Refactor parsing frames
