+++ Runtime/Model/BacktraceRawStackTraceParser.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d450322870c65034cab6c1e809eea6e9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
