+++ Tests/Runtime/Model.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8d214bb14609c3b47937d0fd46e8c1f7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
